+++ examples/next-ts/pages/listbox/basic.tsx
@@ -31,7 +31,7 @@ export default function Page() {
           <label {...api.getLabelProps()}>Label</label>
           <ul {...api.getContentProps()}>
             {selectData.map((item) => (
-              <li key={item.value} {...api.getItemProps({ item })}>
+              <li key={item.value} {...api.getItemProps({ item, highlightOnHover: true })}>
                 <span {...api.getItemTextProps({ item })}>{item.label}</span>
                 <span {...api.getItemIndicatorProps({ item })}>✓</span>
               </li>