+++ packages/machines/password-input/src/password-input.machine.ts
@@ -11,12 +11,6 @@ export const machine = createMachine<PasswordInputSchema>({
       autoComplete: "current-password",
       ignorePasswordManagers: false,
       ...props,
-      translations: {
-        visibilityTrigger(visible) {
-          return visible ? "Hide password" : "Show password"
-        },
-        ...props.translations,
-      },
     }
   },